Make decorator global
