+++ .storybook/preview.tsx
@@ -1,0 +1,33 @@
+import type { Preview } from "@storybook/react";
+
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+import { BrowserRouter } from 'react-router-dom';
+import { ChakraProvider } from '@chakra-ui/react';
+import React from 'react';
+
+const queryClient = new QueryClient();
+
+export const decorators = [
+  (Story) => (
+    <ChakraProvider>
+      <QueryClientProvider client={queryClient}>
+        <BrowserRouter>
+          <Story />
+        </BrowserRouter>
+      </QueryClientProvider>
+    </ChakraProvider>
+  ),
+];
+
+const preview: Preview = {
+  parameters: {
+    controls: {
+      matchers: {
+        color: /(background|color)$/i,
+        date: /Date$/i,
+      },
+    },
+  },
+};
+
+export default preview;
